requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
psycopg2==2.8.6
pyTelegramBotAPI==3.7.9
python-dotenv==0.17.1
requests==2.25.1
urllib3==1.26.5
